Remove nitpick mode from sphinx-build calls
The Django documentation does not use the full paths of classes in the
documentation. For instance, Django docs use django.forms.Modelform
instead of django.forms.models.ModelForm (the actual path). When
generating documentation, Sphinx discovers the latter, but raises an
error with nitpick enabled because the intersphinx tool cannot find the
reference in Django. The solution would be to go through and find every
instance Django does this, or else to remove nitpick. We opt for the
latter.